{{cookiecutter.api_name}}/requirements.txt
ruamel.yaml==0.15.35
requests==2.18.4
